apps/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2
langchain>=0.1.0
openai>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
<<<<<<< HEAD
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
python-multipart>=0.0.5
pytest>=6.2.5
=======
# We're using WooCommerce API instead of direct database access
# WooCommerce API client
woocommerce>=3.0.0
requests>=2.28.0
python-multipart>=0.0.5
pytest>=6.2.5

# Google Calendar API
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# HTTP client
httpx>=0.24.0

# Voice recognition
whisper>=1.0.0

# Email handling
# smtplib is a built-in Python library

# Translation services
translators>=5.4.0
googletrans>=4.0.0-rc1

# File handling
python-magic>=0.4.27

# JWT for authentication
jwt>=1.3.1

# Date and time handling
pytz>=2022.1
python-dateutil>=2.8.2
>>>>>>> 9c26091 (backend try)
